requirements.txt
appdirs
asgiref
certifi
cffi
charset-normalizer
cryptography
Django
django-filter
djangorestframework
openai
djangorestframework-simplejwt
fernet
idna
Markdown
mysqlclient
pyaes
pycparser
PyJWT
pytz
requests
sqlparse
tzdata
urllib3
django-allauth
Pillow
django-rest-auth
djoser
gunicorn
django-heroku
stripe
django-environ
django-mathfilters
